+++ packages/genre-tree-view/src/index.ts
@@ -1,9 +1,11 @@
 export { GenreTree } from "./GenreTree";
 export { GenreTreeWheel } from "./GenreTreeWheel";
 export { GenreTreeWheelRight } from "./GenreTreeWheelRight";
+export { GenreTreeWheelRadial } from "./GenreTreeWheelRadial";
 export { getGenreTreeColor, WHEEL_DEFAULT_FRAME_HEIGHT } from "./constants";
 export { groupNodesByRoot } from "./root-grouping";
 export type { GenreTreeNode, GenreTreeProps, GenreTreePlayState, TreeOrientation } from "./types";
 export type { GenreTreeWheelProps } from "./GenreTreeWheel";
 export type { GenreTreeWheelRightProps } from "./GenreTreeWheelRight";
+export type { GenreTreeWheelRadialProps } from "./GenreTreeWheelRadial";
 export type { GenreTreeRootGroup } from "./root-grouping";